circleci: don't tag to git or deploy to docker hub
Doing all of this in Circle turned out to be a pain, so I'm backing out those bits.

Eventually, it would be good to have all of those various tasks centralized: once we do that elsewhere, we can remove the `deployment` bit entirely here. In the meantime, we should keep the existing behavior.