Add rake task to scrape rubocop documentation
This adds the `docs:scrape` rake task which will pull down the rubocop
repo and scrape all of the cops for class documentation and insert it
into markdown files.

This uses regular expresses to extract and manipulate the class
documentation since working with YARD and RDoc like this isn't
straightforward.