Allow remediation points to be multipliable
When a cop has a score/limit we now multiply a per violation amount of
points by the overage of the violation. eg: 25/20 would multiply the
per violation points by 5

This also changes `config/cops.yml` cops to have two values,
`base_points` and `violation_points`.

Since Rubocop violations don't directly expose the score/limit we check
for the presence and extract the values from the format of
`[score/limit]` in the violation message.