Monkey patch class and method length cops
Rubocop `ClassLength` and `MethodLength` cops have a function called
`check_code_length` that determines the lines to report for the
violation.

Rubocop passes in `keyword` as the method to be called on `node`
which always returns `1` since it's grabbing the defining keyword, eg:
`class` or `def`.

This monkey patch passes in the node's `loc` property directly to return
the correct lines.