Refactor Rubocop runner into more objects
Code Climate triggered a warning that this class was too large and
looking closely it seems to have several different responsibilities
mixed together. At the same time, a lot of the language used in this
engine doesn't match the language of how we talk about our domain.

These changes aim to break down the Rubocop class into collaborators
and to refine the language to bring its concerns closer to our
domain language.

* Introduce SourceFile
* Rename ViolationDecorator to Issue
* Move presentation concerns into Issue
* Refine remediation point names to better match what they represent
* Begin to improve coverage on remediation point generation