Patch RuboCop::Config to warn on obsolete cops
Instead of raising on the inclusion of an obsolete cop, write an error
to STDERR. This is slightly more invisible than failing a build on an
obsolete cop, but preferable as a user won't be blocked when we bump
RuboCop on our engine. We'll write a changelog post advising users of
the necessary changes to make and deprecate this patch in the coming
weeks.