Resolve all paths to absolute paths
When paths were sent in as relative paths from the CLI, `Exclude`
directives within a rubocop config yml were not correctly applied. This
is because Rubocop internally turns those excludes into absolute paths,
and it seems like it was then trying to match the being-analyzed
relative path against the had-been-made-absolute excluded path.