Fix doc scraping for newer rubocop
RuboCop has added magic comments for encoding & frozen string literals
to all of their source, which breaks our attempts to extract doc
comments. This changes the logic to discard all comment lines
at the beginning of the file, and only use the remainder of the file
after the first non-comment line to look for the doc comments.