Commit cop documentation
Directly commits content from cop class documentation that at least
provides an example or a description that isn't just a restatement
of the check name or description.

The motivation behind this change is to ensure that the content being
exposed through the Code Climate service is useful and gives the user
more details on how to resolve the issue being reported rather than
restating a portion of the description. By directly committing the
content this allows us to start supplementing the content from other
sources such as the Ruby styleguide and our own original content.

* Stop scraping content during the Docker image build
* Scrape documentation from the current version of RuboCop
* Modify regular expression to better target class documentation
* Replace good/bad rdoc tokens with comments