Bump RuboCop to 0.47.1 (#89)
The change in the Dockerfile is motivated by this issue and it's
suggested workaround: https://github.com/sickill/rainbow/issues/48

I generated the docs outside of docker because the rake task doesn't
currently work inside docker. I think that's fine.

The `Registry` change is to maintain parity with an internal RuboCop
refactoring that took place as part of this update.