Fix docs:scrape regex to skip forward to class definition
This scraper was missing the documentation on files like
lib/rubocop/cop/style/conditional_assignment.rb
and
lib/rubocop/cop/layout/indent_first_argument.rb
Because they had a comment before the main docs comment. This was
causing them to be rejected because the docs were not long enough,
despite the fact that these files do in fact have docs.

A solution I am proposing for this problem is to use a look ahead to
grab the comments that come immediately before the class definition as
this is where the docs comments always are.